change vpath_to_stc to return a double-precision matrix

+++ utils/general/vpath_to_stc.m
@@ -3,7 +3,7 @@ function Gamma = vpath_to_stc(vpath,K)
 
 if size(vpath,2)>1, error('Viterbi paths have one single column'); end
 if nargin<2, K = max(vpath); end
-Gamma = zeros(numel(vpath),K,'single');
+Gamma = zeros(numel(vpath),K,'double');
 for k=1:K, Gamma(vpath==k,k) = 1; end        
 
 end
